requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
openpyxl>=3.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
scikit-learn>=1.1.0